+++ Example/CoreDataManager.swift
@@ -32,11 +32,11 @@ class CoreDataManager
     fileprivate let applicationDocumentsDirectory = FileManager.default.urls(for :.documentDirectory, in: .userDomainMask).last!
     
     fileprivate let managedObjectModel : NSManagedObjectModel = {
-        let modelURL = Bundle.main.url(forResource: "Banks", withExtension: "momd")
+        let modelURL = Bundle.main.url(forResource: "Banks", withExtension: "momd")!
         #if swift(>=3.2)
-            return NSManagedObjectModel(contentsOf: modelURL!)
+            return NSManagedObjectModel(contentsOf: modelURL)!
         #else
-            return NSManagedObjectModel(contentsOf: modelURL!)!
+            return NSManagedObjectModel(contentsOf: modelURL)!
         #endif
     }()